Master_Paper/model/requirements.txt
numpy
tensorflow-macos
matplotlib
tensorflow-metal
azure-eventhub
azure-storage-blob
python-dotenv